requirements.txt
# Core dependencies
Flask==2.1.2
Flask-Cors==3.0.10
python-dotenv==0.20.0
requests==2.28.1
Pillow==9.2.0
numpy==1.23.3
python-magic==0.4.27

# Image processing
opencv-python==4.6.0.66
scikit-image==0.19.3

aiohttp==3.8.1
beautifulsoup4==4.11.1
lxml==4.9.1

# Development/testing
pytest==7.1.3
pytest-cov==4.0.0

pip install flask-cors dotenv aiohttp feedparser bs4 flask[async] 